working on database
updating id from int to long, doing update, delete and adding methods, connect and disconnect methods and adding prepared boolean for documents if prepared

+++ agnosco-backend/src/main/scala/model/common/Document.scala
@@ -1,4 +1,4 @@
 package model.common
 
 
-case class Document(id : Int, name : String, pages : Iterable[Page], prepared: Boolean)
+case class Document(id : Long, name : String, pages : Iterable[Page], prepared: Boolean)
