+++ agnosco-backend/src/main/scala/model/common/Example.scala
@@ -1,3 +1,3 @@
 package model.common
 
-case class Example(id : Int, imagePath : String, var transcript : Option[String])
+case class Example(id : Long, imagePath : String, var transcript : Option[String])
